test cleanup, testdata re-homing
